added basic ifswitch constant modifier
we're still operating under the "__tmpcosmo_" paradigm,
so the user will need to provide the temporary errno constants. It is
possible to have the compiler handle the generation of the actual
subsituted values, but handling errors with the generation is a pain, so
might as well crash in the usual ways.