adding ifswitch logic into gcc
we call the ifswitch parser rewriting logic right before the
PLUGIN_PRE_GENERICIZE call happens for C files (not sure if we need to
do a similar thing for C++ files at the moment)

If this doesn't work out, an alternative for later is that we could
probably do a --enable-plugin on the GCC configure script so that
ENABLE_PLUGIN is valid, and then add the AST rewriting logic as the
usual plugin-style code.

I didn't do the above right now because I'm not sure it makes sense to
add a plugin callback when the plugin is compiled into GCC. What I think
will happen is that we will add some additional logic into the
invoke_plugin_callbacks so that we can jump into the portcosmo spots.