re-enable C++ support
the reason we had to do this separately was because the function
build_component_ref is not available from the C++ frontend (thereby
triggering an error when linking cc1plus). To handle that, I separated
out the only function that involves build_component_ref, and used its
equivalent (cp_build_addr_expr) in portcosmo_bcref_cp.cc for C++, and now
we get a successful build.