enable AST rewrite WITHOUT using temporaries
earlier, to perform the AST rewriting, we needed temporary constants of
the type __tmpcosmo_EINVAL to be available (either as static const or as
a #define) so that we could later use the temporary values and the
location of the substitution to figure out what actual values to put
there. with this commit, we no longer need to do that! here's what
happens:

let's assume the non-constant error is raised at a switch-case or
initializer, which contains the expression E. here's what we now do:

- if E is a VAR_DECL, we search for it/add it to the hashmap and
  business goes on as earlier
- if E can be simplified to a VAR_DECL, we do the same as above
- if E is a complicated expression, we add a NEW UNIQUE ENTRY to the
  hash table, corresponding to just that constant expression

That's it! when doing the unpatching later, there is a unique temporary
integer assigned to each VAR_DECL and arbitrary expression that we
found, so we use that integer to figure out what to do.

this can still explode for int/short types, but otherwise it seems to be
solid.