patch break statements more than one level deep
we had a simple loop that went over the body of the switch statement,
that looked to replace case labels and break statements with the
appropriate goto statements.

but this was insufficient because some break statements could be hidden
within a BIND_EXPR (a new scope is created in a case), so we wrote a
nested loop to handle BIND_EXPR_BODY.

but this too turned out to be insufficient, because break statements
could be present in the body of an if statement as well (see
src/extract.c in the source code of GNU tar). This means the a
particular case label inside a switch could have nested ifs and
BIND_EXPR and then a valid break at the innermost of all these
statement.

to handle breaks within arbitrary scopes/ifs, this commit changes the
code to perform case label replacement by walking through the body of
the switch statements, and the tree traversal ensures we can reach any
break statement nested inside.