handle Duff's device and similar switch situations
once inside a switch statement S1, case labels can technically be
anywhere, and they belong S1 unless they are inside another switch
statement. so we have to go through the entire AST of the switch to see
if there are any case statements to modify.

once inside a switch statement S1, break statements that are within
other breakable constructs (switch/for/while/do-while) should NOT be
modified, but any break statements within the scope (ie contained within
the toplevel of the switch or within an arbitrary number of ifs/scopes)
should be modified to the right exit label.

hence we walk through the AST twice, once for cases and once for breaks,
so as to be able to handle situations like Duff's device.