remove __attribute__((fallthrough))
switch statements often have an __attribute__((fallthrough)) to avoid
the -Wimplicit-fallthrough warning. when rewriting the switch, we need
to ensure these fallthrough statements are ignored.

note that fallthroughs, like breaks, are only checked within the
top-level of the switch, because, according to GCC online docs,
__attribute__((fallthrough)) "can only be used in a switch statement
(the compiler will issue an error otherwise), after a preceding
statement and before a logically succeeding case label, or user-defined
label."