fix error in initialization of global integers
when the initstruct part of the plugin was originally added to this
patch in 9be0310561c274, we only had the capability to handle
initializations like

        static struct c = {.a = O_DIRECT, .b = EINVAL};

and we did not have the capability to handle patching arbitrary
expressions, which was added much later in commits 98a010e70a65c46
and 14352d9819f672, which meant we could handle initializations like

        static struct c = {.a = O_DIRECT | O_NODELAY, .b = EINVAL};

however, I did not update all parts of initstruct code to benefit from
this new capability. Specifically, I did not update the part of
initstruct that handled the initializations of global _integers_,
although I did update the part for global structs, local structs, and
local ints. (note the code before this commit had 3 calls of the
function arg_should_be_unpatched, not 4).

Thus, a crash was found when trying to run the tests contained within
gnutls-3.7.10, which had a global integer initialization like:

        int a = (O_DIRECT | O_NODELAY);

This commit fixes the error by calling arg_should_be_unpatched at the
necessary spot. A minimal example with the above line was tested, and
now the compiler doesn't segfault.