avoid IDENTIFIER_NAME in struct initialization
applied patch from
https://github.com/ahgamut/gcc/commit/d6b2544e444f53ea9934b4c368628144fc5d1a9b

before this commit, we were using IDENTIFIER_NAME to properly re-create
the constructor for structs that were using magic constants. We were
using this to access the correct member of the structure by name, and
fill in its initial value properly.

However, C structs can contain anonymous unions as members, and we
cannot access the union itself by name, because it doesn't have -- we
can only access the members of the union (which are called
CONSTRUCTOR_ELT or TYPE_FIELDS in the gcc internals docs, depending on
if its an initialization or in general) .

This detail caused an issue with the initialization of struct sigaction,
as seen in https://github.com/jart/cosmopolitan/issues/1183

The current fix is to simply avoid using IDENTIFIER_NAME and use
DECL_NAME instead; we don't need to lookup the struct member by its name
during the reconstruction process, we just need to get its
IDENTIFIER_NODE somehow, because that is what is required by
build_component_ref in the construction. (This should have been
considered by comparing portcosmo_bcref with its C++ version, which is a
oneliner that does not involve any name lookups).

This fix avoids the compiler crash, as confirmed by the example in
https://github.com/ahgamut/portcosmo-examples/commit/147ad1c6e0589ce573b7f57eda651df57ca0ddbf
The compiler successfully patches the struct with the correct values. It
remains to be seen if the compiler can patch magic values added in the
anonymous union.