stub ARROW_LOC parameter in build_component_ref
it appears that the build_component_ref function has added an extra
parameter called ARROW_LOC, to provide a nicer error when there are
issues with pointer-to-a-pointer kind of situations. I don't think
ARROW_LOC is relevant to struct processing performed by this patch,
because the parameter is only used in build_component_ref to show a
nicer error message when the error is triggered by user code. Most
usages of build_component_ref provide UNKNOWN_LOCATION as the input
so, I will do the same unless we find that ARROW_LOC is necessary.